Address @ihhub comments
