+++ src/fheroes2/maps/maps_fileinfo.h
@@ -196,6 +196,10 @@ using MapsFileInfoList = std::vector<Maps::FileInfo>;
 namespace Maps
 {
     MapsFileInfoList PrepareMapsFileInfoList( const bool multi );
+
+    // Creates the list of fheroes2 Resurrection map files.
+    MapsFileInfoList prepareResurrectionMapsFileInfoList();
+
 }
 
 #endif
